Implement compare API.
This implements fetching a diff by pathspec, we need to implement this
upstream so it's implemented as a workaround for now.

This remoes support for copy and rename finding for now. I filed #10849
to fix this properly in a separate step.

The actual diff output format is hard-coded and not yet implemented as
serializers for now which makes experimenting much easier in case we
need to adapt the format depending on how the client implements it.

Fixes #10323