Unpin celery, upgrading it to 4.2.x
This caused a few tests that do manual retrying of a task to fail in
unit tests, celery now throwing an error if you try to fetch a result
within a task, which retry() does. This appears to be only in eager
mode though (works fine locally), so I removed the test and used
celery's autoretry_for feature instead of our manual retry.