Correctly stream the raw file content.
If passed a bytestring as data the streamed content will be bonkers, if
passed a BytesIO file-like object the actual data will be correctly
streamed.

In theory, with a BytesIO wrapping the memoryview we should be keeping
the advantages of using `memoryview` for the blob, which is reduced
memory usage. In practice this probably won't matter much.

I have no idea how to actually test the memory-exhausting to be honest,
so this is primarily guesswork and docs-reading.

Fixes #1118