Add flag for versions needing human reviews, and display them in reviewer tools (#12712)
Add flag for versions needing human reviews, and display them in reviewer tools

This flag is going to be set automatically by scanners. It needs to be set on
versions and not add-ons to ensure specific versions that were found to be
problematic by scanners are examined by human reviewers.