Don't deal with multiple data types at the same time in reindex command
This simplification allows us to fix an issue with the reindex workflow
being wrong with Celery > 4.1.1, which completely broke reindexing of
a single index.

A new --data parameter is introduced for when we want to reindex stats
and not addons. We'll have to fire the command twice if we want to
reindex both.