Iterate through deltas instead of patches to generate entries.
Fixes #13291

This change will use a different pygit2 API to iterate through the
metadata of a diff without accessing it's textual representation.

That'll speed up the generation of `entries` for the compare views.

There might still be some significant time spent generating the diffs
and `entries` as we do fetch some data to figure out correct mimetype
but we should see at least some speedup (according to pygit2 issue
tracker it's significant, unsure how it compares to other places we
spend lots of time on)