Refactor blocklist import task to avoid extra db queries
Because the query to create and then update the kinto import instance
were interrupted by a potentially long query to fetch the guid(s), we
could get a MySQL Server has gone away error as django was trying to
re-use a connection that might have timed out.

Let's try to avoid that by consolidating KintoImport creation into
a single query at the end.