Split translations per locale in Elasticsearch instead of per analyzer
This avoids storing ['foo', 'foo', 'foo'] for 'english' if the add-on
has translations for 'en-US', 'en-CA' and 'en-GB', and in turn prevents
such cases from being considered more relevant than they really are.

To allow for users searching in a language that has multiple variants
in a locale different from what the add-on has provided, we still
search through all variants, but we only consider the best match.