Simplify ActivityLog.objects.for_*() methods and implement for_versions()
The simplification makes us do JOINs instead of an extra query and avoids
a costly evaluation of the results just to build a list of pks when there
are many logs returned.

for_versions() replaces for_version(), works the same way as for_addons(),
if a single Version instance is passed it converts it to a tuple of one
item behind the scenes.