Make reviewer tools pages larger and stop displaying counts in queue tab bar
The queue counts are costly and we can live without them in the tab bar. It's
probably best to hide them rather than cache them and deal with them not
always be up to date.

The width of all reviewer tools pages is also adjusted to make the layout a
little nicer - now that we don't have the counts anymore, 120em should be
enough.