Mark devhub/forms/widgets/compat_app_input_option.html as a django template
Otherwise l10n extraction is broken as it tries to load it as a django template.

Django itself uses settings.FORM_RENDERER or the form default_renderer/renderer
properties, which in the end defaults to the django template loader, so it should
be considered as a django template at all times.