Add Accept-Language to Vary headers in the API responses
When there is no lang in the URL, the activated language for the request
will depend on Accept-Language header, so we need to Vary on it in case we did
return translated content (not necessarily translations, could also be URLs
that contain a locale prefix) depending on language.

Requests outside the API could technically have the same problem, but for 99%
of them LocaleAndAppURLMiddleware will force a redirect to a locale-aware URL
(adding correct vary header to the redirect), sidestepping the issue.