Remove handling of X_IS_MOBILE_AGENTS and 404 responsive template (#16949)
The 404 responsive was a bit of a hack introduced when we were in the process
of integrating the new frontend, it lacked links to make it really useful and
its design was completely different from the rest, because users weren't
supposed to see it all that much: it was a fallback needed because we had both
the legacy frontend and the new frontend, and users could switch from one to
the other.

It has never been improved since. and nowadays it's only used when we had a 404
in devhub/reviewer tools/admin from a User-Agent detected as being a mobile
one, which is rare. It's not really worth keeping as-is, it forces us to Vary
on User-Agent in addons-server, which we'd like to avoid, and depends on a
header sent by nginx we'd like to stop supporting.