Handle License.name using our existing translation fields
Because of builtin licenses and the way they have their names translated using
gettext(), we were previously handling License.name in a completely custom way,
emulating the behavior of TranslationField, but we didn't update that for v5.

This makes it use the existing fields so that it will automatically follow
their behavior in the future.