Fix reviewers review log filtering to include force disable/enable for listed reviewers
Flip the logic around: if the user doesn't have unlisted permission, instead of
filtering to only show activities linked to a version on listed channel, exclude those
on unlisted channel. That makes activities not linked to a particular version, like
force disable/enable, visible for those users.

Same idea applied to the review link URL: instead of defaulting to unlisted and changing
that if the activity is linked to a listed version, default to listed and only link to
unlisted review page if the activity is linked to an unlisted version.