Stop filtering ES queries on doc type
doc type is deprecated and will eventually be removed. We need to stop passing
it at index creation in the mapping and at indexing time, and stop passing it
when searching.

Because we are still currently passing it to search queries, if we were to
remove it right away from index creation/bulk indexing, search would break and
never find any results. So the first step is to remove it from search queries,
and then later remove it from everywhere else too.