Find abuse reports for a particular add-on using guid and not addon_id
When an abuse report is submitted from a guid (what Firefox does),
no longer try to associate if with a single add-on in the database,
because we can't know which one it is.

In reviewer tools, look up abuse reports for a given add-on by guid,
so that we can find all potential abuse reports even if a guid has
been re-used.