Prevent request.user from being evaluated in middlewares
When dealing with a request with auth credentials, GraphiteMiddleware's
default implementation forces request.user to be evaluated to send a
`response.auth` statsd ping, causing a database query even if the view didn't
use auth at all. We don't need that ping, so replace the middleware with a
custom one that doesn't do this, avoiding the database query in those cases.