Reload version in auto_approve before determining auto-approval verdict
The general idea is to improve consistency checks and avoid accidentally
auto-approving an add-on that isn't in the right state for it to happen.

In order to do that, on top of loading the version at the last possible time,
a refactor of ReviewHelper was necessary to make non human reviews more
explicit and check add-on status before making the approval action available,
which allows auto_approve to check for action availability instead of
bypassing that check as it did before.