Only consider extensions in the `mad`/`scanner` reviewer queues and scanner actions
This allows the query to look at fewer rows, making it faster - the scanner
actions also avoid touching non-extensions to make things consistent and avoid
the situation where something would trigger an action we wouldn't notice
because it's not shown in the queues.

In addition, in the reviewer queues, the COUNT() is optimized using the same
method already used for dashboard counts, by not returning columns we won't
need and avoiding sorting in the subquery.