Remove M2M between ScannerQueryResult and ScannerQueryRule, use FK instead
For queries, there should be one rule per result only. This allows deletion
of query rules to properly delete query results, and in turn avoids the
issue where orphaned results would stay around and mess up filtering when
only one rule was remaining.

Because it's in the way, also move things specific to ScannerResult and
ScannerRule that ScannerQueryResult/ScannerQueryRule are not using outside
of the common base classes.