Set fields with an expensive default when initializing models in ESAddonSerializer
When a model __init__() is called, fields are set to their default values,
which can sometimes have a significant cost when they are callables. This
is notably the case for UserProfile.auth_id, and to a lesser extend created.

While we're at it also provide a cheaper default for auth_id in tests.