Stop doing phrase matching for single-term queries in search
Also stop decompounding the word "all" to avoid decompounding "wallet" and
"paywall". In most names where "all" is present with no whitespace (already a
rare occurrence), it's at the end (so our prefix matching is enough) and/or
with a word we are already decompounding anyway (download, browser, tab, etc).