Add word-break:break-word to uses of white-space:pre-wrap
`white-space: pre-wrap` prevents lines from being broken unless an
explicit line break is in there.

Since the content of these fields are external (user-generated), there
is no guarantee of the line length to be within reasonable bounds.

Rather than stretching the layout, add `word-break: break-word` to allow
the line to be wrapped regardless.