Commit the transaction earlier when reporting/appealing to Cinder
This should help ensure the `CinderJob` exists as soon as possible,
when we are reporting a user with a ton of relationships. This is
necessary because we need that `CinderJob` to exist to process
decisions, so if a moderator is acting super fast on a report but
we're still busy sending relationships and haven't committed the
transaction yet, we can't execute the decision.