Update Reply-To in SHA1 resign script and send mails separately (#22175)
* Update Reply-To in SHA1 resign script and send mails separately

Operations asked to change the Reply-To.

The emails were grouped by add-on, which would potentially leak
the emails of other owners of the add-on when they are more than
one. That information is already available in the authors page for
the add-on in devhub but since the email is not encrypted it's
best to avoid that problem entirely by sending separate emails.