Clear NeedsHumanReview selectively depending on job being resolved (#22401)
If there are multiple jobs for a given add-on, when resolving a job we only want to clear the NeedsHumanReview flag if the job we're resolving is the last one of its kind: the add-on should stay in the queue until all jobs are resolved.

With that logic in place, because reports can be added to an ongoing appeal job, we need to also prevent NeedsHumanReview from being added following a report attached to an appeal job (otherwise we can't determine what NeedsHumanReview to clear). It's redundant anyway, there will already have a NeedsHumanReview for the appeal and that's enough.