Remove obsolete validator caching
This was only really relevant when validation could be triggered arbitrarily by
reviewers for an existing File(Upload) that was laying around unvalidated.

That confusing and dangerous functionality no longer exists, FileUpload
validation is always done once at upload and can't be triggered later.

When we still had post-request-task, that caching was broken for years
and nobody noticed because the code path where we pulled the task id
from the cache was never used.