Stop using our remote settings "test server mode" for dev, it's broken
It's been broken for ages - that mode was designed to work around the fact
that the Kinto server was being regularly wiped out, so we had this special
user that created the user and collection dynamically, and then used its
own special bucket. But nowadays we can just use the same
user/bucket/collection everywhere, as the remote settigns server has the
same config everywhere.