Change growth threshold algorithm do use percentage points increase
Also contains a refactor to make the code easier to test, as well as
admin & reviewer tools tweaks to make the consequences of changing that
threshold more obvious.

If the average hotness in a tier is 0.1, and the growth threshold is set to 50,
we used to flag add-ons in that tier with hotness over 0.15 (percentage increase)
now we'll flag over 0.51 (percentage points increase).