Record `reasoning` and `private_notes` separately in Content Decisions (#23419)
* Record `reasoning` and `private_notes` separately in Content Decisions

- `reasoning` is the comment sent to the content owner (or reporter for
  a reporter appeal).
- `private_notes` are, as the name implies, private notes entered by
  the moderator, and not shared externally.

For consistency and to keep existing data, since we had already shared
existing `notes` externally, the `notes` field has been renamed to
`reasoning` and a new `private_notes` field has been introduced.

When private notes are entered for a decision that goes through
ContentAction.log_action(), record a separate activity for it,
mimic what we do in reviewer tools in that case.

* Include private notes in stakeholder email

* Document use of reasoning for reviewer comments