Fix potential IntegrityError exception in EmailUserRestriction admin
Since we're normalizing patterns on save (and when checking at upload) if
someone tried to record a non-normalized variant of an email that we already
had in the database, it would raise an IntegrityError exception.

This fixes that problem by normalizing earlier, so that we raise a proper
validation error to the user.

Slight drive-by: remove some unnecessary gettext calls around
restrictions-related strings that only show up in the admin.