Avoid raising exceptions when parsing invalid manifests (#24185)
* Avoid raising exceptions when parsing invalid manifests

Validate that fields that should contain a dict do return
one - we're calling this parsing code before the linter has
run, so the manifest could be invalid.

When we detect such invalid fields, raise InvalidManifest,
the validation task catches that to let the linter run and
provide an appropriate error message to the developer.

* Test fixes

* Use Mapping as the type to check, default to empty dict fallback where needed